start nicegui server when main_file is not provided
