+++ nicegui/testing/screen_plugin.py
@@ -9,10 +9,12 @@ import pytest
 from selenium import webdriver
 from selenium.webdriver.chrome.service import Service
 
+from ..ui_run import run
 from .general_fixtures import (  # noqa: F401  # pylint: disable=unused-import
+    get_path_to_main_file,
     nicegui_reset_globals,
+    prepare_simulation,
     pytest_configure,
-    resolve_main_path,
 )
 from .screen import Screen
 
@@ -75,8 +77,7 @@ def screen(nicegui_reset_globals,  # noqa: F811, pylint: disable=unused-argument
            ) -> Generator[Screen, None, None]:
     """Create a new SeleniumScreen fixture."""
     screen_ = Screen(nicegui_driver, caplog)
-    main_path = resolve_main_path(request.config)
-    screen_.server_thread = threading.Thread(target=runpy.run_path, args=(main_path,))
+    screen_.server_thread = create_server(request)
     screen_.server_thread.start()
     yield screen_
     logs = [record for record in screen_.caplog.get_records('call') if record.levelname == 'ERROR']
@@ -87,3 +88,11 @@ def screen(nicegui_reset_globals,  # noqa: F811, pylint: disable=unused-argument
         shutil.rmtree(DOWNLOAD_DIR)
     if logs:
         pytest.fail('There were unexpected ERROR logs.', pytrace=False)
+
+
+def create_server(request: pytest.FixtureRequest) -> threading.Thread:
+    main_path = get_path_to_main_file(request.config)
+    if main_path is None:
+        prepare_simulation()
+        return threading.Thread(target=run)
+    return threading.Thread(target=runpy.run_path, args=(main_path,))
