+++ nicegui/testing/user_plugin.py
@@ -6,16 +6,16 @@ from typing import Callable
 import httpx
 import pytest
 
-import nicegui
 from nicegui import core, ui
 from nicegui.functions.download import download
 from nicegui.functions.navigate import Navigate
 from nicegui.functions.notify import notify
 
 from .general_fixtures import (  # noqa: F401  # pylint: disable=unused-import
+    get_path_to_main_file,
     nicegui_reset_globals,
+    prepare_simulation,
     pytest_configure,
-    resolve_main_path,
 )
 from .user import User
 
@@ -29,7 +29,7 @@ async def user(nicegui_reset_globals,  # noqa: F811, pylint: disable=unused-argu
                ) -> AsyncGenerator[User, None]:
     """Create a new user fixture."""
     os.environ['NICEGUI_USER_SIMULATION'] = 'true'
-    runpy.run_path(resolve_main_path(request.config))
+    runpy.run_path(get_path_to_main_file(request.config))
     async with core.app.router.lifespan_context(core.app):
         async with httpx.AsyncClient(transport=httpx.ASGITransport(core.app), base_url='http://test') as client:
             yield User(client)
@@ -51,23 +51,3 @@ async def create_user(nicegui_reset_globals) -> AsyncGenerator[Callable[[], User
     ui.navigate = Navigate()
     ui.notify = notify
     ui.download = download
-
-
-def prepare_simulation() -> None:
-    """Prepare a simulation to be started.
-    """
-    core.app.config.add_run_config(
-        reload=False,
-        title='Test App',
-        viewport='',
-        favicon=None,
-        dark=False,
-        language='en-US',
-        binding_refresh_interval=0.1,
-        reconnect_timeout=3.0,
-        message_history_length=1000,
-        tailwind=True,
-        prod_js=True,
-        show_welcome_message=False,
-    )
-    nicegui.storage.set_storage_secret('simulated secret')
